dbeaver/cloudbeaver#663 :fire: Removing shell scripts that calls deploy/docker
As docker-compose and Dockerfile are usually dropped in the root dir
because of the context, developers should be able to just build
and run containers from the root dir. That way, it's similar
to most open-source projects.

docker-compose build
docker-compose up

As an interface, users don't need to keep moving around to different
directories in order to build and deploy a server locally.

Also, the container will run in a localhost context, so
any network connectivity should address routing to a host.
For this reason, the script deploy/docker/run-docker-container.sh
doesn't make sense to have.