CB-5999 fix: don't see changes when change connection view
The root of this problem for flat files connections in the fact the on the top level in a connection, we get the same node, so we don't go deeper, when use refreshNode and children remain their previous state. I didn't find a way to detect this case somehow based on results of queries, so I just used refreshTree method for changing connection view, which updates all the children.