Add mantine popover
Popover.Target requires an element or a component as a single child – strings, fragments, numbers and multiple elements/components are not supported and will throw error. Custom components must provide a prop to get root element ref. It could be a problem for the cloudbeaver codebase

popover mounted in wrong place