CB-6346 refactor: change condition for column read-only indicator
Previously we used to check dataReadonly variable and hide indicator on the columns if it's true, because dataReadonly was true only when all columns were in read-only mode. Currently dataReadonly doesn't depend on columns' read-only status and may be false even in cases when all the columns are read-only. But we know for sure that if we don't have an elements' identifier in the table, we consider the table read-only and all the columns will be in read-only mode. This new check returns the previous behavior back in new conditions.